makes StLambda with catching exceptions (same as TrLambda)
